pyproject.toml
[project]
name = "docflow"
version = "0.1.0"
description = "Automated document scanning, tagging, and archiving for macOS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.4",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "anthropic>=0.28.0",
    "apscheduler>=3.10.4",
    "structlog>=24.2.0",
    "pillow>=10.3.0",
    "img2pdf>=0.5.1",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
macos = [
    "osxphotos>=0.68.0",
    "pyobjc-framework-Vision>=10.3",
    "pyobjc-framework-Cocoa>=10.3",
    "pyobjc-core>=10.3",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
docflow = "docflow.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docflow"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "unit: fast unit tests (no I/O, no network)",
    "e2e: end-to-end tests (real SQLite, real files, mocked external APIs)",
]

[tool.uv]
environments = ["sys_platform == 'darwin' and platform_release >= '22.0'"]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.11"
warn_return_any = false
ignore_missing_imports = true
